packages/boot-azure-function/pyproject.toml
[project]
name = "alt-python-boot-azure-function"
version = "1.0.0"
description = "Azure Function adapter for alt-python/boot"
requires-python = ">=3.12"
dependencies = [
    "alt-python-boot-lib",
    "azure-functions>=1.18",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["boot_azure_function"]
